requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
cffi==2.0.0
click==8.3.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.0.1
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
